split tws suite
